Redefine uint32_t as "unsigned long"
This is so converting from a pointer to a uint32_t becomes easier.

+++ stdint.h
@@ -3,9 +3,9 @@
 
 typedef unsigned char          uint8_t;
 typedef unsigned short         uint16_t;
-typedef unsigned int           uint32_t;
+typedef unsigned long          uint32_t;
 typedef unsigned long long int uint64_t;
 
-typedef unsigned long uintptr_t;
+typedef uint32_t uintptr_t;
 
 #endif
